Update to 4.5.4 nativecontrolsfortouch and Schema fixes
git-svn-id: https://plugins.svn.wordpress.org/video-embed-thumbnail-generator/tags/4.5.4@1216612 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ video-embed-thumbnail-generator.php
@@ -3,10 +3,11 @@
 Plugin Name: Video Embed & Thumbnail Generator
 Plugin URI: http://www.kylegilman.net/2011/01/18/video-embed-thumbnail-generator-wordpress-plugin/
 Description: Generates thumbnails, HTML5-compliant videos, and embed codes for locally hosted videos. Requires FFMPEG or LIBAV for encoding.
-Version: 4.5
+Version: 4.5.4
 Author: Kyle Gilman
 Author URI: http://www.kylegilman.net/
 Text Domain: video-embed-thumbnail-generator
+Domain Path: /languages
 
 This program is free software; you can redistribute it and/or modify
 it under the terms of the GNU General Public License as published by
@@ -58,7 +59,7 @@ function kgvid_default_options_fn() {
 	$edit_others_capable = kgvid_check_if_capable('edit_others_posts');
 
 	$options = array(
-		"version" => 4.5,
+		"version" => 4.504,
 		"embed_method" => "Video.js",
 		"jw_player_id" => "",
 		"template" => false,
@@ -268,7 +269,12 @@ function kgvid_get_attachment_meta($post_id) {
 				delete_post_meta($post_id, "_kgflashmediaplayer-".$key);
 			}
 
-			array_filter($kgvid_postmeta, 'strlen'); //remove empty elements
+			foreach ( $kgvid_postmeta as $key => $value ) {
+				if ( $value === null ) {
+					unset( $kgvid_postmeta[ $meta ] ); //remove empty elements
+				}
+			}
+
 			kgvid_save_attachment_meta($post_id, $kgvid_postmeta);
 
 		}
@@ -283,9 +289,11 @@ function kgvid_get_attachment_meta($post_id) {
 
 function kgvid_save_attachment_meta($post_id, $kgvid_postmeta) {
 
-	$options = kgvid_get_options();
-
 	if ( is_array($kgvid_postmeta) ) {
+
+		$options = kgvid_get_options();
+		$kgvid_old_postmeta = kgvid_get_attachment_meta($post_id);
+		$kgvid_postmeta = array_merge($kgvid_old_postmeta, $kgvid_postmeta); //make sure all keys are saved
 
 		foreach ( $kgvid_postmeta as $key => $meta ) { //don't save if it's the same as the default values or empty
 
@@ -1441,7 +1449,7 @@ function kgvid_video_embed_enqueue_scripts() {
 
 	//Video.js styles
 	if ( $options['embed_method'] == "Video.js" || $options['embed_method'] == "Strobe Media Playback" ) {
-		wp_enqueue_style( 'video-js', plugins_url("", __FILE__).'/video-js/video-js.css', '', '4.12.7' );
+		wp_enqueue_style( 'video-js', plugins_url("", __FILE__).'/video-js/video-js.css', '', '4.12.11' );
 		if ( $options['js_skin'] == 'kg-video-js-skin' ){ wp_enqueue_style( 'video-js-kg-skin', plugins_url("", __FILE__).'/video-js/kg-video-js-skin.css', '', $options['version'] ); }
 	}
 
@@ -1625,7 +1633,7 @@ function kgvid_enqueue_shortcode_scripts() {
 
 	if ( $options['embed_method'] == "Video.js" || $options['embed_method'] == "Strobe Media Playback" ) {
 			wp_enqueue_script( 'video-quality-selector', plugins_url("", __FILE__).'/video-js/video-quality-selector.js', array('video-js'), $options['version'], true );
-			wp_enqueue_script( 'video-js', plugins_url("", __FILE__).'/video-js/video.js', '', '4.12.7', true );
+			wp_enqueue_script( 'video-js', plugins_url("", __FILE__).'/video-js/video.js', '', '4.12.11', true );
 			add_action('wp_footer', 'kgvid_print_videojs_footer', 99);
 		}
 
@@ -1979,8 +1987,16 @@ function KGVID_shortcode($atts, $content = ''){
 
 				if ( !empty($id) ) { //if the video is an attachment in the WordPress db
 
-					$content = wp_get_attachment_url($id);
-					if ( $content == false ) { echo "Invalid video ID"; continue; }
+					$attachment_url = wp_get_attachment_url($id);
+					if ( $attachment_url == false ) { echo "Invalid video ID"; continue; }
+					$exempt_cdns = array('amazonaws.com', 'rackspace.com', 'netdna-cdn.com', 'nexcess-cdn.net', 'limelight.com'); //don't replace URLs that point to CDNs
+					$exempt_url = false;
+					foreach ( $exempt_cdns as $exempt_cdn ) {
+						if ( strpos($content, $exempt_cdn) !== false ) {
+							$exempt_url = true;
+						}
+					}
+					if ( !$exempt_url ) { $content = $attachment_url; }
 
 					$encodevideo_info = kgvid_encodevideo_info($content, $id);
 					$attachment_info = get_post( $id );
@@ -1997,7 +2013,7 @@ function KGVID_shortcode($atts, $content = ''){
 					if ( !empty($poster_id) ) {
 						$poster_image_src = wp_get_attachment_image_src($poster_id, 'full');
 						$query_atts['poster'] = $poster_image_src[0];
-						if ( intval($query_atts['width']) <= get_option('medium_size_h') ) {
+						if ( strpos($query_atts['width'], '%') === false && intval($query_atts['width']) <= get_option('medium_size_h') ) {
 							$query_atts['poster'] = kgvid_get_attachment_medium_url($poster_id);
 						}
 					}
@@ -2106,8 +2122,7 @@ function KGVID_shortcode($atts, $content = ''){
 
 				} //if Strobe Media
 
-				//wp_localize_script( 'kgvid_video_embed', 'kgvid_video_vars_'.$div_suffix, $video_variables ); //add video variables in footer
-
+				$video_variables = str_replace('&nbsp;', ' ', $video_variables);
 				$code .= '<div id="kgvid_'.$div_suffix.'_wrapper" class="kgvid_wrapper'.$aligncode.'">'."\n\t\t\t";
 				$code .= '<div id="video_'.$div_suffix.'_div" class="fitvidsignore kgvid_videodiv" data-id="'.$div_suffix.'" data-kgvid_video_vars="'.esc_attr(json_encode($video_variables)).'" itemprop="video" itemscope itemtype="https://schema.org/VideoObject">';
 				if ( $query_atts["poster"] != '' ) { $code .= '<meta itemprop="thumbnailUrl" content="'.esc_attr($query_atts["poster"]).'" />'; }
@@ -2117,10 +2132,17 @@ function KGVID_shortcode($atts, $content = ''){
 				$code .= '<meta itemprop="contentUrl" content="'.$content.'" />';
 
 				if ( !empty($query_atts['title']) ) { $code .= '<meta itemprop="name" content="'.esc_attr($query_atts['title']).'" />'; }
+
 				if ( !empty($query_atts['description']) ) { $description = $query_atts['description']; }
 				elseif ( !empty($query_atts['caption']) ) { $description = $query_atts['caption']; }
-				else { $description = ""; }
-				if ( !empty($description) ) { $code .= '<meta itemprop="description" content="'.esc_attr($description).'" />'; }
+				elseif ( in_the_loop() && !is_attachment() ) { $description = get_the_excerpt(); }
+				if ( empty($description) ) { $description = __('Video', 'video-embed-thumbnail-generator'); }
+				$code .= '<meta itemprop="description" content="'.esc_attr($description).'" />';
+
+				if ( !empty($id) ) { $upload_date = get_the_date('c', $id); }
+				elseif ( $post_ID != 1 ) { $upload_date = get_the_date('c', $post_ID); }
+				else { $upload_date = current_time('c'); }
+				$code .= '<meta itemprop="uploadDate" content="'.esc_attr($upload_date).'" />';
 
 				$track_keys = array('kind', 'srclang', 'src', 'label', 'default');
 				if ( !isset($kgvid_postmeta) || ( is_array($kgvid_postmeta) && !is_array($kgvid_postmeta['track']) ) ) {
@@ -2138,7 +2160,7 @@ function KGVID_shortcode($atts, $content = ''){
 							if ( empty($value) ) { $track_attribute[$attribute] = $query_atts['track_'.$attribute]; }
 						}
 						if ( $options['embed_method'] == "WordPress Default" && $track_attribute['kind'] == 'captions' ) { $track_attribute['kind'] = 'subtitles'; }
-						$track_code .= "\t\t\t\t\t<track kind='".esc_attr($track_attribute['kind'])."' src='".esc_attr($track_attribute['src'])."' srclang='".esc_attr($track_attribute['srclang'])."' label='".esc_attr($track_attribute['label'])."' ".esc_attr($track_attribute['default'])." />\n";
+						$track_code .= "\t\t\t\t\t<track id='".$div_suffix."_text_".$track."' kind='".esc_attr($track_attribute['kind'])."' src='".esc_attr($track_attribute['src'])."' srclang='".esc_attr($track_attribute['srclang'])."' label='".esc_attr($track_attribute['label'])."' ".esc_attr($track_attribute['default'])." />\n";
 					}
 				}
 
@@ -2294,7 +2316,9 @@ function KGVID_shortcode($atts, $content = ''){
 					if ( is_array($atts) && array_key_exists('skin', $atts) ) {
 						$options['js_skin'] = $atts['skin']; //allows user to set skin for individual videos using the skin="" attribute
 					}
-					$code .= ' class="fitvidsignore '.esc_attr('video-js '.$options['js_skin']).'" data-setup=\'{ "nativeControlsForTouch": true, "language": "'.substr($locale, 0, 2).'"';
+					$code .= ' class="fitvidsignore '.esc_attr('video-js '.$options['js_skin']).'" data-setup=\'{ ';
+					if ( $options['nativecontrolsfortouch'] == "on" ) { $code .= '"nativeControlsForTouch": true, '; }
+					$code .= '"language": "'.substr($locale, 0, 2).'"';
 					if ( $enable_resolutions_plugin ) { $code .= ', "plugins" : { "resolutionSelector" : { "force_types" : ["video/mp4"]';
 						if ( $query_atts["auto_res"] == "highest" ) { $code .= ', "default_res": "'.end($h264_resolutions).'"'; }
 						if ( $query_atts["auto_res"] == "lowest" ) { $code .= ', "default_res": "'.reset($h264_resolutions).'"'; }
@@ -3312,6 +3336,7 @@ function kgvid_video_embed_options_init() {
 	add_settings_field('audio', __('Volume:', 'video-embed-thumbnail-generator'), 'kgvid_audio_callback', __FILE__, 'kgvid_video_embed_playback_settings', array( 'label_for' => 'volume' ) );
 	add_settings_field('preload', __('Preload:', 'video-embed-thumbnail-generator'), 'kgvid_preload_callback', __FILE__, 'kgvid_video_embed_playback_settings', array( 'label_for' => 'preload' ) );
 	add_settings_field('js_skin', _x('Skin class:', 'CSS class for video skin', 'video-embed-thumbnail-generator'), 'kgvid_js_skin_callback', __FILE__, 'kgvid_video_embed_playback_settings', array( 'label_for' => 'js_skin' ) );
+	add_settings_field('nativecontrolsfortouch', __('Native controls:', 'video-embed-thumbnail-generator'), 'kgvid_nativecontrolsfortouch_callback', __FILE__, 'kgvid_video_embed_playback_settings', array( 'label_for' => 'nativecontrolsfortouch' ) );
 	add_settings_field('custom_attributes', __('Custom attributes:', 'video-embed-thumbnail-generator'), 'kgvid_custom_attributes_callback', __FILE__, 'kgvid_video_embed_playback_settings', array( 'label_for' => 'custom_attributes' ) );
 
 	add_settings_field('bgcolor', __('Background color:', 'video-embed-thumbnail-generator'), 'kgvid_bgcolor_callback', __FILE__, 'kgvid_video_embed_flash_settings', array( 'label_for' => 'bgcolor' ) );
@@ -3574,6 +3599,11 @@ add_action('admin_init', 'kgvid_video_embed_options_init' );
 	function kgvid_js_skin_callback() {
 		$options = kgvid_get_options();
 		echo "<input class='regular-text code affects_player' id='js_skin' name='kgvid_video_embed_options[js_skin]' type='text' value='".$options['js_skin']."' /><br /><em><small>".sprintf( __('Use %s for a nice, circular play button. Leave blank for the default square play button.', 'video-embed-thumbnail-generator')." <a href='http://videojs.com/docs/skins/'>".__('Or build your own CSS skin.', 'video-embed-thumbnail-generator'), '<code>kg-video-js-skin</code>')."</a></small></em>\n\t";
+	}
+
+	function kgvid_nativecontrolsfortouch_callback() {
+		$options = kgvid_get_options();
+		echo "<input class='affects_player' ".checked( $options['nativecontrolsfortouch'], "on", false )." id='nativecontrolsfortouch' name='kgvid_video_embed_options[nativecontrolsfortouch]' type='checkbox' /> <label for='nativecontrolsfortouch'>".__('Show native controls on mobile devices.', 'video-embed-thumbnail-generator')."</label> <a class='kgvid_tooltip wp-ui-text-highlight' href='javascript:void(0);'><span class='kgvid_tooltip_classic'>".__('Native controls allow for streaming services like AirPlay and Chromecast, but don\'t always look as nice.', 'video-embed-thumbnail-generator')."</span></a>\n\t";
 	}
 
 	function kgvid_custom_attributes_callback() {
@@ -4256,6 +4286,10 @@ function kgvid_update_settings() {
 			$options['ffmpeg_old_rotation'] = "on";
 			$options['click_download'] = "on";
 		}
+		if ( $options['version'] < 4.504 ) {
+			$options['version'] = 4.504;
+			$options['nativecontrolsfortouch'] = "on";
+		}
 
 		if ( $options['version'] != $default_options['version'] ) { $options['version'] = $default_options['version']; }
 		if ( $options !== $options_old ) { update_option('kgvid_video_embed_options', $options); }
@@ -4864,13 +4898,15 @@ function kgvid_image_attachment_fields_to_edit($form_fields, $post) {
 				}
 				$track_type_select .= "</select>";
 
+				if ( !array_key_exists( 'default', $kgvid_postmeta['track'][$track] ) ) { $kgvid_postmeta['track'][$track]['default'] = false; }
+
 				$tracks_html .= '<div id="kgflashmediaplayer-'.$post->ID.'-trackdiv-'.$track.'" class="kgvid_thumbnail_box kgvid_track_box"><strong>'._x('Track', 'captions track', 'video-embed-thumbnail-generator').' '.strval($track+1).'</strong><span class="kgvid_track_box_removeable" onclick="jQuery(this).parent().remove();jQuery(\'form.compat-item input\').first().change();">X</span><br />
 				'.__('Track type:', 'video-embed-thumbnail-generator').' '.$track_type_select.'<br />
 				<span id="pick-track'.$track.'" class="button-secondary" style="margin:10px 0;" data-choose="'.__('Choose a Text File', 'video-embed-thumbnail-generator').'" data-update="'.__('Set as track source', 'video-embed-thumbnail-generator').'" data-change="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_src" onclick="kgvid_pick_attachment(this);">'.__('Choose from Library', 'video-embed-thumbnail-generator').'</span><br />
 				URL: <input name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][src]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_src" type="text" value="'.$kgvid_postmeta['track'][$track]['src'].'" class="text" style="width:180px;"><br />
 				'._x('Language code:', 'two-letter code indicating track\'s language', 'video-embed-thumbnail-generator').' <input name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][srclang]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_srclang" type="text" value="'.$kgvid_postmeta['track'][$track]['srclang'].'" maxlength="2" style="width:40px;"><br />
 				'.__('Label:', 'video-embed-thumbnail-generator').' <input name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][label]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_label" type="text" value="'.$kgvid_postmeta['track'][$track]['label'].'" class="text" style="width:172px;"><br />
-				'.__('Default:', 'video-embed-thumbnail-generator').'<input '.checked(array_key_exists('default', $kgvid_postmeta['track'][$track]), true, false).' name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][default]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_default" type="checkbox" value="default"></div>';
+				'.__('Default:', 'video-embed-thumbnail-generator').'<input '.checked($kgvid_postmeta['track'][$track]['default'], 'default', false).' name="attachments['. $post->ID .'][kgflashmediaplayer-track]['.$track.'][default]" id="attachments-'. $post->ID .'-kgflashmediaplayer-track_'.$track.'_default" type="checkbox" value="default"></div>';
 			}
 		}
 
@@ -5259,11 +5295,26 @@ function kgvid_video_attachment_fields_to_save($post, $attachment) {
 					update_post_meta($post['ID'], '_kgflashmediaplayer-poster-id', $thumb_id);
 				}
 			}
-			else {
+
+			if ( empty($thumb_url) ) {
 				delete_post_meta($post['ID'], '_kgflashmediaplayer-poster');
 				delete_post_meta($post['ID'], '_kgflashmediaplayer-poster-id');
+				delete_post_thumbnail($post['ID']);
 			}
-			update_post_meta($post['ID'], '_kgflashmediaplayer-poster', $thumb_url);
+			else {
+
+				update_post_meta($post['ID'], '_kgflashmediaplayer-poster', $thumb_url);
+
+				if ( empty($thumb_id) ) { //we're not saving a new thumbnail
+					$poster_id = get_post_meta($post['ID'], '_kgflashmediaplayer-poster-id', true);
+					if ( empty($poster_id) ) { //the poster_id meta was accidentally deleted
+						$thumb_url_id = kgvid_url_to_id($thumb_url);
+						if ( $thumb_url_id ) { update_post_meta($post['ID'], '_kgflashmediaplayer-poster-id', $thumb_url_id); }
+					}
+				}
+
+			}
+
 		}
 
 		if( isset($attachment['kgflashmediaplayer-featured']) ) {
@@ -5299,6 +5350,12 @@ function kgvid_video_attachment_fields_to_save($post, $attachment) {
 		$checkboxes = array( 'lockaspect', 'featured', 'showtitle', 'downloadlink' ); //make sure unchecked checkbox values are saved
 		foreach ( $checkboxes as $checkbox ) {
 			if( !isset($attachment['kgflashmediaplayer-'.$checkbox]) ) { $attachment['kgflashmediaplayer-'.$checkbox] = "false"; }
+		}
+
+		if ( isset( $attachment['kgflashmediaplayer-track'] ) && is_array( $attachment['kgflashmediaplayer-track'] ) ) {
+			foreach ( $attachment['kgflashmediaplayer-track'] as $track => $track_attribute ) {
+				if ( !array_key_exists('default', $track_attribute) ) { $attachment['kgflashmediaplayer-track'][ $track ]['default'] = ''; }
+			}
 		}
 
 		$kgvid_postmeta = array();
@@ -5593,7 +5650,7 @@ function kgvid_video_attachment_template() {
 	if ( $options['embeddable'] == 'false' && !array_key_exists('sample', $kgvid_video_embed) && !array_key_exists('gallery', $kgvid_video_embed) ) { $kgvid_video_embed['enable'] = 'false'; }
 
 	if ( is_array($kgvid_video_embed) && array_key_exists('enable', $kgvid_video_embed) && $kgvid_video_embed['enable'] == 'true'
-		&& ( ( array_key_exists('post_mime_type', $post) && strpos($post->post_mime_type, 'video') !== false ) || array_key_exists('sample', $kgvid_video_embed) )
+		&& ( ( is_object($post) && array_key_exists('post_mime_type', $post) && strpos($post->post_mime_type, 'video') !== false ) || array_key_exists('sample', $kgvid_video_embed) )
 		) {
 
 		$html = kgvid_generate_embeddable_video( $kgvid_video_embed );
